Progress (#18)
* Add renderer constants

* GPU buffer progress

* Sprite progress

* Draw 2D sprites with different textures

* Address the `uint` infection

* Update submodules

* Update `flatc` EXEs to allow build

* Integrate `yaLanTingLibs` library

* Remove `FlatBuffers` library

* Update CMakeLists.txt

* Don't use general-purpose parallel executor for renderer thread

* Use `std::thread` instead of `std::jthread` for potential WebASM compatibilityat

* Renderer cleanup

* Minor changes

* Minor changes

* Combine 2D primitive and 2D sprite shaders

* Simpler texture access, minor revisions

* Renderer progress

* Renderer progress

* Update Renderer.cpp

* Beginnings of a batching system

* Renderer progress

* Add mutex locks

* Organise

* Organise, draw batch progress

* Store 2D shape triangles in same GPU buffer as 2D sprites

* Organise

+++ Source/Renderer/Backends/SdlGpu/Pipeline.cpp
@@ -112,7 +112,7 @@ namespace Silent::Renderer
         SDL_ReleaseGPUShader(_device, fragShader);
     }
 
-    SDL_GPUShader* PipelineManager::LoadShader(const std::string& filename, uint samplerCount, uint storageTexCount, uint storageBufferCount, uint uniBufferCount)
+    SDL_GPUShader* PipelineManager::LoadShader(const std::string& filename, int samplerCount, int storageTexCount, int storageBufferCount, int uniBufferCount)
     {
         // Define shader stage.
         auto stage = SDL_GPUShaderStage::SDL_GPU_SHADERSTAGE_VERTEX;
@@ -178,10 +178,10 @@ namespace Silent::Renderer
             .entrypoint           = entryPoint,
             .format               = activeFormatFlag,
             .stage                = stage,
-            .num_samplers         = samplerCount,
-            .num_storage_textures = storageTexCount,
-            .num_storage_buffers  = storageBufferCount,
-            .num_uniform_buffers  = uniBufferCount
+            .num_samplers         = (uint)samplerCount,
+            .num_storage_textures = (uint)storageTexCount,
+            .num_storage_buffers  = (uint)storageBufferCount,
+            .num_uniform_buffers  = (uint)uniBufferCount
         };
 
         // Create shader.
